feat: single main function - refactored away code duplication
