+++ src/sys/linux/mod.rs
@@ -1,3 +1,4 @@
 pub mod cgroup;
 pub mod mem_info;
 pub mod allocator;
+pub mod system;
